discord_bot/requirements.txt
aiohttp
beautifulsoup4
discord.py
loguru
pydantic
python-dotenv
Requests
stem
sqlalchemy
mdformat
mdformat-gfm